refactor: unify navigraph login (#8119)
* Initial commit

* Make unlink account actually clear persistent properties

* Fix JS module build scripts to know about all env vars

* Add forgotten files

* Add override simbrief id functionality

* Change loc keys for userid validation

* Add option to delete override user id

* Formatting

* Minor fixes and tweaks. NEW LOC KEY

* Fix build scripts dotenv config working dir

* Support .env.local, fix source map config

* Fix atsu common build

* fix: old URL object being reused

---------
