+++ fbw-a32nx/src/base/flybywire-aircraft-a320-neo/html_ui/Pages/A32NX_Utils/SimBriefApi.js
@@ -1,7 +1,11 @@
+// Copyright (c) 2021-2023 FlyByWire Simulations
+//
+// SPDX-License-Identifier: GPL-3.0
+
 class SimBriefApi {
-    static getSimBriefOfp(userId) {
-        if (userId) {
-            return fetch(`${SimBriefApi.url}&userid=${userId}`)
+    static getSimBriefOfp(username, overrideUserID) {
+        if (username || overrideUserID) {
+            return fetch(overrideUserID ? `${SimBriefApi.url}&userid=${overrideUserID}` : `${SimBriefApi.url}&username=${username}`)
                 .then((response) => {
                     if (!response.ok) {
                         throw new HttpError(response.status);
@@ -10,7 +14,7 @@ class SimBriefApi {
                     return response.json();
                 });
         } else {
-            throw new Error("No SimBrief pilot ID provided");
+            throw new Error("No Navigraph username or override simbrief user id provided");
         }
     }
 }
