+++ fbw-a32nx/src/systems/atsu/fmsclient/build.js
@@ -6,29 +6,8 @@
 
 const esbuild = require('esbuild');
 const path = require('path');
+const { esbuildModuleBuild } = require('#build-utils');
 
-const rootDir = path.join(__dirname, '..', '..', '..');
-const outFile = '../out/flybywire-aircraft-a320-neo/html_ui/JS/fbw-a32nx/atsu/fmsclient.js';
+const outFile = 'fbw-a32nx/out/flybywire-aircraft-a320-neo/html_ui/JS/fbw-a32nx/atsu/fmsclient.js';
 
-const isProductionBuild = process.env.A32NX_PRODUCTION_BUILD === '1';
-
-esbuild.build({
-    absWorkingDir: __dirname,
-
-    define: { DEBUG: 'false' },
-
-    entryPoints: ['src/index.ts'],
-    bundle: true,
-    treeShaking: false,
-    minify: isProductionBuild,
-
-    outfile: path.join(rootDir, outFile),
-
-    format: 'iife',
-    globalName: 'AtsuFmsClient',
-
-    sourcemap: isProductionBuild ? 'linked' : undefined,
-
-    // Target approximate CoherentGT WebKit version
-    target: 'safari11',
-});
+esbuild.build(esbuildModuleBuild('fbw-a32nx', 'AtsuFmsClient', path.join(__dirname, 'src/index.ts'), outFile));
