+++ fbw-a32nx/src/systems/instruments/src/EFB/Dashboard/Widgets/FlightWidget.tsx
@@ -68,7 +68,8 @@ const NoSimBriefDataOverlay = ({ simbriefDataLoaded, simbriefDataPending, fetchD
 export const FlightWidget = () => {
     const { data } = useAppSelector((state) => state.simbrief);
     const [simbriefDataPending, setSimbriefDataPending] = useState(false);
-    const [simbriefUserId] = usePersistentProperty('CONFIG_SIMBRIEF_USERID');
+    const [navigraphUsername] = usePersistentProperty('NAVIGRAPH_USERNAME');
+    const [overrideSimBriefUserID] = usePersistentProperty('CONFIG_OVERRIDE_SIMBRIEF_USERID');
     const [airframe] = useState(getAirframeType());
 
     const {
@@ -114,7 +115,7 @@ export const FlightWidget = () => {
         setSimbriefDataPending(true);
 
         try {
-            const action = await fetchSimbriefDataAction(simbriefUserId ?? '');
+            const action = await fetchSimbriefDataAction(navigraphUsername ?? '', overrideSimBriefUserID ?? '');
 
             dispatch(action);
         } catch (e) {
