+++ fbw-a32nx/src/systems/instruments/src/EFB/Settings/Settings.tsx
@@ -1,3 +1,7 @@
+// Copyright (c) 2021-2023 FlyByWire Simulations
+//
+// SPDX-License-Identifier: GPL-3.0
+
 import React, { FC } from 'react';
 
 import { Route, Switch } from 'react-router';
@@ -85,11 +89,29 @@ export const SettingsPage: FC<SettingsPageProps> = ({ name, backRoute, children 
             </div>
         </Link>
         <div className="py-2 px-6 w-full rounded-lg border-2 h-content-section-reduced border-theme-accent">
-            <ScrollableContainer height={53}>
+            <ScrollableContainer height={54} innerClassName="h-full">
                 <div className="h-full divide-y-2 divide-theme-accent">
                     {children}
                 </div>
             </ScrollableContainer>
+        </div>
+    </div>
+);
+
+export const FullscreenSettingsPage: FC<SettingsPageProps> = ({ name, children }) => (
+    <div>
+        <Link to="/settings" className="inline-block mb-4">
+            <div className="flex flex-row items-center space-x-3 hover:text-theme-highlight transition duration-100">
+                <ArrowLeft size={30} />
+                <h1 className="font-bold text-current">
+                    {t('Settings.Title')}
+                    {' - '}
+                    {name}
+                </h1>
+            </div>
+        </Link>
+        <div className="py-2 px-6 w-full h-content-section-reduced rounded-lg border-2 border-theme-accent">
+            {children}
         </div>
     </div>
 );
