+++ fbw-a32nx/src/systems/systems-host/build.js
@@ -6,28 +6,8 @@
 
 const esbuild = require('esbuild');
 const path = require('path');
+const { esbuildModuleBuild } = require('#build-utils');
 
-const rootDir = path.join(__dirname, '..', '..', '..');
-const outFile = 'out/flybywire-aircraft-a320-neo/html_ui/Pages/VCockpit/Instruments/A32NX/SystemsHost';
+const outFile = 'fbw-a32nx/out/flybywire-aircraft-a320-neo/html_ui/Pages/VCockpit/Instruments/A32NX/SystemsHost/index.js';
 
-const isProductionBuild = process.env.A32NX_PRODUCTION_BUILD === '1';
-
-esbuild.build({
-    absWorkingDir: __dirname,
-
-    define: { DEBUG: 'false' },
-
-    entryPoints: ['./index.ts'],
-    bundle: true,
-    treeShaking: false,
-    minify: isProductionBuild,
-
-    outdir: path.join(rootDir, outFile),
-
-    format: 'iife',
-
-    sourcemap: isProductionBuild ? 'linked' : undefined,
-
-    // Target approximate CoherentGT WebKit version
-    target: 'safari11',
-});
+esbuild.build(esbuildModuleBuild('fbw-a32nx', undefined, path.join(__dirname, './index.ts'), outFile));
