+++ fbw-common/src/systems/shared/src/index.ts
@@ -1,3 +1,4 @@
+export * from './navigraph';
 export * from './ApproachUtils';
 export * from './arinc429';
 export * from './ata';
